+++ plugins/tags-resources/src/utils.ts
@@ -1,9 +1,10 @@
 // Copyright © 2022 Hardcore Engineering Inc.
 
 import { Doc, FindResult, Ref } from '@anticrm/core'
-import { createQuery, LiveQuery } from '@anticrm/presentation'
-import { TagElement, TagReference } from '@anticrm/tags'
+import { TagReference } from '@anticrm/tags'
 import tags from './plugin'
+import { FilterQuery } from '@anticrm/view-resources'
+import { Filter } from '@anticrm/view'
 
 export function getTagStyle (color: string, selected = false): string {
   return `
@@ -12,44 +13,24 @@ export function getTagStyle (color: string, selected = false): string {
   `
 }
 
-export const TagFilterQuery = {
-  queries: new Map<number, LiveQuery>(),
-  results: new Map<number, Array<Ref<Doc>>>(),
-
-  getLiveQuery (index: number): LiveQuery {
-    const current = TagFilterQuery.queries.get(index)
-    if (current !== undefined) return current
-    const query = createQuery(true)
-    this.queries.set(index, query)
-    return query
-  },
-
-  async getRefs (res: Array<Ref<TagElement>>, onUpdate: () => void, index: number): Promise<Array<Ref<Doc>>> {
-    const lq = this.getLiveQuery(index)
-    const promise = new Promise<Array<Ref<Doc>>>((resolve, reject) => {
-      const refresh = lq.query(
-        tags.class.TagReference,
-        {
-          tag: { $in: res }
-        },
-        (refs: FindResult<TagReference>) => {
-          const result = Array.from(new Set(refs.map((p) => p.attachedTo)))
-          TagFilterQuery.results.set(index, result)
-          resolve(result)
-          onUpdate()
-        }
-      )
-      if (!refresh) {
-        resolve(TagFilterQuery.results.get(index) ?? [])
+export async function getRefs (filter: Filter, onUpdate: () => void): Promise<Array<Ref<Doc>>> {
+  const lq = FilterQuery.getLiveQuery(filter.index)
+  const promise = new Promise<Array<Ref<Doc>>>((resolve, reject) => {
+    const refresh = lq.query(
+      tags.class.TagReference,
+      {
+        tag: { $in: filter.value }
+      },
+      (refs: FindResult<TagReference>) => {
+        const result = Array.from(new Set(refs.map((p) => p.attachedTo)))
+        FilterQuery.results.set(filter.index, result)
+        resolve(result)
+        onUpdate()
       }
-    })
-    return await promise
-  },
-
-  remove (index: number): void {
-    const lq = this.queries.get(index)
-    lq?.unsubscribe()
-    this.queries.delete(index)
-    this.results.delete(index)
-  }
+    )
+    if (!refresh) {
+      resolve(FilterQuery.results.get(filter.index) ?? [])
+    }
+  })
+  return await promise
 }